+++ new_actors.c
@@ -572,7 +572,7 @@ void add_enhanced_actor_from_server(char * in_data)
 	if(strlen(&in_data[28]) >= 30)
 		{
 			char str[120];
-			snprintf(str, 120, "Bad actor name/length (%d): %s/%d\n", actors_list[i]->actor_type,&in_data[28], strlen(&in_data[28]));
+			snprintf(str, 120, "Bad actor name/length (%d): %s/%d\n", actors_list[i]->actor_type,&in_data[28], (int)strlen(&in_data[28]));
 			log_error(str);
 			return;
 		}
